// examples/playground/src/zero/schema.gen.ts
// @generated drizzle-zero signature:sha256:bda40fa32438f777793af44491d47c283a5cda124e41084c1492a9a290176863
// This file was automatically generated by drizzle-zero.
// You should NOT make any changes in this file as it will be overwritten.

import type { Row } from "@rocicorp/zero";
import { createBuilder } from "@rocicorp/zero";

const issueTable = {
    "name": "issue",
    "columns": {
        "id": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string
        },
        "title": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string
        },
        "completed": {
            "type": "boolean",
            "optional": false,
            "customType": null as unknown as boolean
        },
        "ownerId": {
            "type": "string",
            "optional": false,
            "customType": null as unknown as string
        },
        "createdAt": {
            "type": "number",
            "optional": false,
            "customType": null as unknown as number
        }
    },
    "primaryKey": ["id"]
} as const;
/**
 * The Zero schema object.
 * This type is auto-generated from your Drizzle schema definition.
 */
export const schema = {
    "tables": {
        "issue": issueTable
    },
    "relationships": {},
    "enableLegacyQueries": false,
    "enableLegacyMutators": false
} as const;

/**
 * Represents the Zero schema type.
 * This type is auto-generated from your Drizzle schema definition.
 */
export type Schema = typeof schema;
/**
 * Represents a row from the "issue" table.
 * This type is auto-generated from your Drizzle schema definition.
 */
export type Issue = Row<(typeof schema)["tables"]["issue"]>;

/**
 * Represents the ZQL query builder.
 * This type is auto-generated from your Drizzle schema definition.
 */
export const zql = createBuilder(schema);
/**
 * Represents the Zero schema query builder.
 * This type is auto-generated from your Drizzle schema definition.
 */
export const builder = zql;

/** Defines the default types for Zero */
declare module '@rocicorp/zero' { interface DefaultTypes { schema: Schema; } }